Removing duplicated files
In the other examples "Run.m" is the run file so I've renamed your "MainScript" as "Run" and your "Run" and "Run_fn". Four files are duplicates of what's in the code folder so I've removed them and added the code folder to the path directory. A disadvantage of adding the other folder to the directory is that one line differs between Windows and other operating systems, so I've got the non-Windows line commented out for now.